changes to convert.str
