+++ plugins/plugin_loader.py
@@ -23,7 +23,7 @@ class PluginLoader():
     @staticmethod
     def get_converter(name):
         """ Return requested converter plugin """
-        return PluginLoader._import("Convert", "Convert_{0}".format(name))
+        return PluginLoader._import("convert", name)
 
     @staticmethod
     def get_model(name):
@@ -55,6 +55,16 @@ class PluginLoader():
                         if not item.name.startswith("_")
                         and item.name.endswith(".py"))
         return models
+        
+    @staticmethod
+    def get_available_converters():
+        """ Return a list of available converters """
+        converter_path = os.path.join(os.path.dirname(__file__), "convert")
+        converters = sorted(item.name.replace(".py", "").replace("_", "-")
+                        for item in os.scandir(converter_path)
+                        if not item.name.startswith("_")
+                        and item.name.endswith(".py"))
+        return converters
 
     @staticmethod
     def get_available_extractors(extractor_type):
